move share lib and tv-common(zzz_generated/tv-app/zap-generated) link to example, so that we cloud have different android tv apps for different clusters
