+++ BUILD.gn
@@ -125,6 +125,7 @@ if (current_toolchain != "${dir_pw_toolchain}/default:default") {
 
     if (current_os == "android") {
       deps += [
+        "${chip_root}/examples/tv-app/android:java",
         "${chip_root}/src/app/server/java",
         "${chip_root}/src/controller/java",
         "${chip_root}/src/platform/android:java",
