+++ examples/tv-app/android/BUILD.gn
@@ -15,10 +15,15 @@
 import("//build_overrides/build.gni")
 import("//build_overrides/chip.gni")
 
+import("${build_root}/config/android_abi.gni")
+import("${chip_root}/build/chip/java/rules.gni")
 import("${chip_root}/build/chip/tools.gni")
 
-# Todo: copy from examples/tv-app/linux but later will be moved to jni and upto java workd
-source_set("android-tv-app") {
+# Todo: copy from examples/tv-app/linux but later will be moved to jni and upto java word
+shared_library("jni") {
+
+  output_name = "libTvApp"
+  
   sources = [
     "${chip_root}/examples/tv-app/tv-common/include/CHIPProjectAppConfig.h",
     "include/account-login/AccountLoginManager.cpp",
@@ -49,13 +54,46 @@ source_set("android-tv-app") {
     "include/tv-channel/TvChannelManager.h",
     "include/wake-on-lan/WakeOnLanManager.cpp",
     "include/wake-on-lan/WakeOnLanManager.h",
+    "java/TVApp-JNI.cpp",
   ]
 
   deps = [
     "${chip_root}/examples/tv-app/tv-common",
     "${chip_root}/src/lib",
     "${chip_root}/third_party/inipp",
+    "${chip_root}/src/app/server/java:jni",
   ]
 
   cflags = [ "-Wconversion" ]
+
+  output_dir = "${root_out_dir}/lib/jni/${android_abi}"
+
+  ldflags = [ "-Wl,--gc-sections" ]
+}
+
+android_library("java") {
+  output_name = "TvApp.jar"
+
+  deps = [
+    ":android",
+    "${chip_root}/third_party/android_deps:annotation",
+  ]
+
+  data_deps = [
+    ":jni",
+    "${chip_root}/build/chip/java:shared_cpplib",
+  ]
+
+  sources = [
+    "java/src/com/tcl/tvapp/TvApp.java",
+  ]
+
+  javac_flags = [ "-Xlint:deprecation" ]
+
+  # TODO: add classpath support (we likely need to add something like
+  #  ..../platforms/android-21/android.jar to access BLE items)
+}
+
+java_prebuilt("android") {
+  jar_path = "${android_sdk_root}/platforms/android-21/android.jar"
 }
